restaurant_management/requirements.txt
Django==5.1.2
djangorestframework
django-phonenumber-field
psycopg2-binary
python-dateutil
pytest
pytest-django
pytest-mockito
factory-boy
injector
pymemcache
djangorestframework-simplejwt
django-ratelimit
drf-yasg
